XMLParser is now a specialization of JSONParser
